requirements.txt
djangorestframework-jwt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
